feat: add openApi

+++ user-service/src/main/java/ru/users/userservice/service/UserService.java
@@ -1,11 +1,11 @@
 package ru.users.userservice.service;
 
 import jakarta.validation.constraints.Positive;
-import ru.users.userservice.dto.NewUserDto;
-import ru.users.userservice.dto.UpdateUserDto;
-import ru.users.userservice.dto.UserDto;
 import jakarta.validation.Valid;
 import org.springframework.data.domain.PageRequest;
+import ru.users.userservice.model.NewUserDto;
+import ru.users.userservice.model.UpdateUserDto;
+import ru.users.userservice.model.UserDto;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -13,11 +13,11 @@ import java.util.List;
 public interface UserService {
     List<UserDto> getUsers(String name, String surname, LocalDate registrationDate, PageRequest pageRequest);
 
-    UserDto getById(Integer userId);
+    UserDto getById(Long userId);
 
     UserDto add(@Valid NewUserDto dto);
 
-    UpdateUserDto update(@Positive Integer userId, @Valid UpdateUserDto dto);
+    UpdateUserDto update(@Positive Long userId, @Valid UpdateUserDto dto);
 
-    Boolean removeById(@Positive Integer userId);
+    Boolean removeById(@Positive Long userId);
 }
